feat: EditBoard 기능 구현 및 style 변경 (#45)
