+++ src/components/BoardDetail/BoardDetailContent.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-
 import {
   StImg,
   StContent,
@@ -9,8 +8,8 @@ import {
   StAllContentBox,
 } from './styles';
 import { MdDeleteForever } from 'react-icons/md';
-import { FaGithub } from 'react-icons/fa';
-import { useNavigate, useParams } from 'react-router-dom';
+import { FaEdit, FaGithub } from 'react-icons/fa';
+import { Link, useNavigate, useParams } from 'react-router-dom';
 import { auth, db } from 'config/firebase';
 import { deleteDoc, doc, getDoc } from 'firebase/firestore';
 
@@ -47,6 +46,11 @@ export default function BoardDetailContent() {
       </StAllContent>
       <StBtn>
         <FaGithub size="30" />
+        {authEmail === data.userid ? (
+          <Link to={`/boards/${id}/edit`}>
+            <FaEdit size="28" />
+          </Link>
+        ) : null}
         <MdDeleteForever size="30" onClick={() => clickDelete()} />
       </StBtn>
     </StAllContentBox>
