+++ src/pages/EditBoard.jsx
@@ -1,0 +1,6 @@
+import EditBoardContent from 'components/EditBoard/EditBoardContent';
+import React from 'react';
+
+export default function EditBoard() {
+  return <EditBoardContent />;
+}
