+++ src/pages/index.js
@@ -4,5 +4,14 @@ import BoardDetail from './BoardDetail';
 import MyPage from './MyPage';
 import NewBoard from './NewBoard';
 import ProtectedRoute from './ProtectedRoute';
+import EditBoard from './EditBoard';
 
-export { Home, AllBoard, BoardDetail, NewBoard, MyPage, ProtectedRoute };
+export {
+  Home,
+  AllBoard,
+  BoardDetail,
+  NewBoard,
+  MyPage,
+  ProtectedRoute,
+  EditBoard,
+};
